refactor : 엔티티 타입형 객체 타입으로 변경 (#47)
