+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/domain/bakeryfavorite/service/BakeryFavoritePageService.java
@@ -5,11 +5,11 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.breaditnow.common.util.GeoPoint;
 import com.breaditnow.customer.domain.bakeryfavorite.controller.req.GeoPointRequest;
 import com.breaditnow.customer.domain.bakeryfavorite.controller.res.BakeryFavoritePageResponse;
 import com.breaditnow.domain.domain.favorite.dto.BakeryFavoriteDto;
 import com.breaditnow.domain.domain.favorite.repository.customerbakeryfavorite.CustomerBakeryFavoriteRepository;
+import com.breaditnow.domain.domain.vo.GeoPoint;
 
 import lombok.RequiredArgsConstructor;
