+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/domain/productfavorite/service/ProductFavoritePageService.java
@@ -4,10 +4,10 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.breaditnow.common.util.GeoPoint;
 import com.breaditnow.customer.domain.bakeryfavorite.controller.req.GeoPointRequest;
 import com.breaditnow.customer.domain.productfavorite.controller.res.ProductFavoritePageResponse;
 import com.breaditnow.domain.domain.favorite.repository.customerproductfavorite.CustomerProductFavoriteRepository;
+import com.breaditnow.domain.domain.vo.GeoPoint;
 
 import lombok.RequiredArgsConstructor;
